pyproject.toml
[project]
name = "daily-collections"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "antlr4-python3-runtime>=4.13.2",
    "antlr4-tools>=0.2.2",
    "chardet>=7.4.0.post1",
    "duckdb>=1.5.1",
    "factory-boy>=3.3.3",
    "faker>=40.11.1",
    "fastapi>=0.135.2",
    "hiredis>=3.3.1",
    "ipython>=9.10.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "mmh3>=5.2.1",
    "mysqlclient>=2.2.8",
    "numpy>=2.4.3",
    "objprint>=0.3.0",
    "openpyxl>=3.1.5",
    "orjson>=3.11.7",
    "pandas>=3.0.1",
    "paramiko>=4.0.0",
    "pendulum>=3.2.0",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "redis>=7.4.0",
    "requests>=2.33.0",
    "six>=1.17.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
    "sqlacodegen>=4.0.3",
]

[tool.ruff]
line-length = 120
